Change customer balance type from DOUBLE to DECIMAL(10, 2) for improved precision in financial calculations

+++ backend/models/Customer.js
@@ -37,7 +37,7 @@ const Customer = db.define('customer', {
         allowNull: false
     }
     ,balance: {
-        type: DataTypes.DOUBLE,
+        type: DataTypes.DECIMAL(10, 2),
         allowNull: false
     }
 });
